move generateAlbumCover to sepatate file
